Streamlit_app/requirements.txt
streamlit
mysql-connector-python
pandas
python-dotenv
bcrypt